css:itinerary 페이지 wireframe 수정
