+++ client/src/Components/itinerary/PlaceInputBox.js
@@ -12,6 +12,7 @@ const InputWrapper = styled.div`
   width: 100%;
   height: 40px;
   align-items: center;
+  justify-content: space-between;
 
   div {
     width: 100%;
